03_29_2021
Note: An attempt to separate a received message on the side of the second server is proving troublesome; I need to separate the concatenated components of the message as well as decrypt those segments with their appropriate keys. In doing so, I will become closer to extracting the ultimate timestamp which is required to authorize the client's attempted access to the final server. As before, the fifth and six steps of the authentication process require attention in wrapping up this project.,